requirements.txt
asgiref==3.3.4
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
Django==3.2.4
django-ninja==0.13.2
docutils==0.17.1
idna==2.10
importlib-metadata==4.5.0
jeepney==0.6.0
keyring==23.0.1
packaging==20.9
pkginfo==1.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.0
twine==3.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
webencodings==0.5.1
zipp==3.4.1
